awk: fix buggy handling of new -- ranges.
The problem is that when records appear in the middle
of the range which again match from-expr, they
get suppressed.

* share/txr/stdlib/awk.tl (sys:awk-mac-let): Get rid of the
flag-mid variable. It cannot work because middle is a state
in its own right that cannot be inferred from the existing
states (nil, t, :end) and the value of from-expr. We get rid
of the flag and introduce a :mid state value.