bugfix: two issues in mappend* and append*.
Two bugs in these functions, both attributable to
the lazy_appendv implementation:

They destructively catenate the input lists,
much like nconc, even though documented as
non-destructive.

If any input list is infinite, other than the
last input list, that list is forced, resulting
in an infinite loop.

* lib.c (lazy_appendv_func): Rewritten to use a
different algorithm which earnestly allocates a
new lazy cons for each element of the output
sequence, except for the tail part corresponding
to the last list.
(lazy_appendv): Set up the lazy cons according
to the new representation and just return it.
No searching for the tail of the nonempty list,
and no destructive manipulation.