bugfix: code-walk the result of a declined macro.
When a form is a macro, but the macro declines to expand the
form, the form must still be code walked; we can't just return
it and be done.  For instance if it is a function call, its
argument expressions have to be expanded.

This also causes undefined function warning to be generated
properly if a macro declines, and the resulting form is
a function call to a not-yet-defined function.

* eval.c (do_expand): If expand_macro yields the original
form, branch backwards to re-execute the whole if statement
again. Use the fact that the local variable macro is now
non-nil to skip the macro expansion case.