bugfix: intern public symbols in autoload files.
This commit addresses a bug of the following type:

  (defpackage p (:fallback usr))
  (in-package p)
  (let ((lb (new list-builder)))
    lb.(add 1)) ;; problem: this add is p:add

the add symbol is not on the auto-load list; it occurs
as a slot of the list-builder class, but is not interned
when TXR starts. Thus the (:fallback usr) doesn't pick it up.
Expected behavior is that add is usr:add.

* lisplib.c (intern_only): New static function.
(sock_set_entries, build_set_entries, getopts_set_entries,
stream_wrap_set_entries): Define additional lists of
supplemenary symbols which are passed to intern_only
just to be interned in the usr package without autoload
registratiions. These symbols are all slots documented for
public use in various structures defined by the respective
modules managed by these autoload functions.