packages: drop no-fallback-list interning restriction.
Removing the restriction that qualified pkg:sym syntax may not
cause interning to take place if pkg has a nonempty fallback
list. This serves no purpose, only hindering the flexibility
of the package system.

* parser.y (sym_helper): When processing a qualified symbol,
if the package exists, just intern it.

* txr.1: Revise all text which touched on the removed rule,
to remove all traces of it from the documentation.