genvmop: fix broken script.
This broke when I moved asm.tl into the library directory
and set it up for auto-load.

* genvmop.txr: We must not include "asm" any more.
But then there is no auto-load on any other feature of the
assembler other than the class name. The @(mdo)
directive comes in handy; at expansion time we can trigger
auto-load by doing a lookup on the sys:assembler
struct name.