vm: introduce sframe instruction.
This is for allocating a new frame purely on the stack.  The
frame will not be captured by lexical closures, and so can
only be used for non-shared variables and additional
compiler-generated temporaries (if registers run out, for
instance).

* share/txr/stdlib/asm.tl (op-sframe, sframe): New opcode
class and opcode.

* vm.c (vm_do_frame): New static function for the common
implementation of frame and sframe.
(vm_frame): Now just a call with vm_do_frame, passing the flag
indicating that closure capture is enabled for this
environment frame.
(vm_sframe): New static function.

* vmop.h: Regenerated.