regression: excess args not diagnosed.
This was caused by the recent work to reduce consing in
generic_funcall.

* lib.c (generic_funcall): Correct test for too many
arguments. Because we don't normalize the argument list to the
exact number of fixed args, but to at least the fixed args,
the excess args can possibly be part of the fill rather than
part of the list.