compiler: append-each: bugfix for non-lists.
* share/txr/stdlib/compiler.tl: If a non-list sequence like a
string or vector is being accumulated, we must have an append
call, because the accumulation cell is never extended with new
cells. After we replace the cdr of ,accum with a string or
vector atom, (last ,accum) just finds ,accum and so then the
cdr will be clobbered on the next iteration.